requirements.txt
pytest>=8.3.4
allure-pytest>=2.13.5
selenium>=4.27.1
